+++ chartlets.js/CHANGES.md
@@ -29,6 +29,9 @@
   dark, light, and system mode.
 
 * Changed the yet unused descriptor type `CbFunction` for callback functions.
+  - using `schema` instead of `type` property for callback arguments
+  - using `return` object with `schema` property for callback return values
+
 
 ## Version 0.0.29 (from 2024/11/26)